Add tree
This adds a tree data structure, it has been sitting here unstaged since
I created this repo and I've been meaning to fix it up and add it. This
accomplishes one of those things (and it's not fixing it), so this is in
no way guaranteed to work or be 'correct'.